+++ arrow-udf-macros/src/parse.rs
@@ -36,9 +36,10 @@ impl Parse for FunctionAttr {
             .split_once('(')
             .ok_or_else(|| Error::new_spanned(&sig, "expected '('"))?;
         let args = args.trim_start().trim_end_matches([')', ' ']);
-        let (is_table_function, ret) = match ret.trim_start().strip_prefix("setof") {
-            Some(s) => (true, s),
-            None => (false, ret),
+        let (is_table_function, ret) = match ret.trim_start() {
+            s if s.starts_with("setof") => (true, &s[5..]), // -> setof
+            s if s.starts_with(">") => (true, &s[1..]),     // ->>
+            _ => (false, ret),
         };
         parsed.name = name.trim().to_string();
         parsed.args = if args.is_empty() {